refactor: 디렉토리명 대문자 제거
